Revert " Making autoincremented "id" column of "pages" table to be long (JET_coltyp.Currency) instead of int32 (JET_coltyp.Long) to avoid possible "Auto-increment counter has reached maximum value" esent error" This _partially_ reverts commit 3ac1f6c10154847d11166b324ca048023797ca2c - From08To09.cs file was not removed, it has been commented out only
Revert "Making autoincremented "id" column of "usage" table to be long (JET_coltyp.Currency) instead of int32 (JET_coltyp.Long) to avoid  "Auto-increment counter has reached maximum value" esent error"
This _partially_ reverts commit e4c5c5f56bf4b24791e795dc0456a4aebb081050 - From07To08.cs file was not removed, it has been commented out only

+++ Raven.Database/FileSystem/Plugins/AbstractFilePutTrigger.cs
@@ -42,7 +42,7 @@ namespace Raven.Database.FileSystem.Plugins
         {
         }
 
-        public virtual void OnUpload(string name, RavenJObject metadata, long pageId, int pagePositionInFile, int pageSize)
+        public virtual void OnUpload(string name, RavenJObject metadata, int pageId, int pagePositionInFile, int pageSize)
         {
         }
